test(kernel): share KernelHandle test setup
