+++ crates/librefang-kernel/tests/kernel_handle_contract_broader.rs
@@ -1,30 +1,8 @@
-use librefang_kernel::LibreFangKernel;
 use librefang_kernel_handle::KernelHandle;
-use librefang_types::config::KernelConfig;
 
-fn boot() -> (LibreFangKernel, tempfile::TempDir) {
-    let tmp = tempfile::tempdir().expect("failed to create temp directory");
-    let home_dir = tmp.path().to_path_buf();
-    let data_dir = home_dir.join("data");
-    std::fs::create_dir_all(&data_dir).expect("failed to create data directory");
-    std::fs::create_dir_all(home_dir.join("skills")).unwrap();
-    std::fs::create_dir_all(home_dir.join("workspaces").join("agents")).unwrap();
-    std::fs::create_dir_all(home_dir.join("workspaces").join("hands")).unwrap();
+mod common;
 
-    let config = KernelConfig {
-        home_dir,
-        data_dir: data_dir.clone(),
-        network_enabled: false,
-        memory: librefang_types::config::MemoryConfig {
-            sqlite_path: Some(data_dir.join("test.db")),
-            ..Default::default()
-        },
-        ..KernelConfig::default()
-    };
-
-    let kernel = LibreFangKernel::boot_with_config(config).expect("failed to boot test kernel");
-    (kernel, tmp)
-}
+use common::boot_kernel as boot;
 
 #[test]
 fn test_roster_roundtrip() {
