+++ crates/librefang-kernel/tests/kernel_handle_contract_cron_spawn.rs
@@ -1,30 +1,8 @@
-use librefang_kernel::LibreFangKernel;
 use librefang_kernel_handle::KernelHandle;
-use librefang_types::config::KernelConfig;
 
-fn boot() -> (LibreFangKernel, tempfile::TempDir) {
-    let tmp = tempfile::tempdir().expect("failed to create temp directory");
-    let home_dir = tmp.path().to_path_buf();
-    let data_dir = home_dir.join("data");
-    std::fs::create_dir_all(&data_dir).expect("failed to create data directory");
-    std::fs::create_dir_all(home_dir.join("skills")).unwrap();
-    std::fs::create_dir_all(home_dir.join("workspaces").join("agents")).unwrap();
-    std::fs::create_dir_all(home_dir.join("workspaces").join("hands")).unwrap();
+mod common;
 
-    let config = KernelConfig {
-        home_dir,
-        data_dir: data_dir.clone(),
-        network_enabled: false,
-        memory: librefang_types::config::MemoryConfig {
-            sqlite_path: Some(data_dir.join("test.db")),
-            ..Default::default()
-        },
-        ..KernelConfig::default()
-    };
-
-    let kernel = LibreFangKernel::boot_with_config(config).expect("failed to boot test kernel");
-    (kernel, tmp)
-}
+use common::boot_kernel as boot;
 
 fn minimal_manifest() -> &'static str {
     r#"
@@ -129,17 +107,14 @@ async fn test_spawn_agent_returns_valid_identity() {
     assert_eq!(found.name, "test-agent");
 }
 
-#[test]
-fn test_list_agents_returns_manifest_metadata() {
+#[tokio::test(flavor = "multi_thread")]
+async fn test_list_agents_returns_manifest_metadata() {
     let (kernel, _tmp) = boot();
     let kh: &dyn KernelHandle = &kernel;
 
-    let rt = tokio::runtime::Builder::new_multi_thread()
-        .enable_all()
-        .build()
-        .unwrap();
-    let (id, _name) = rt
-        .block_on(kh.spawn_agent(minimal_manifest(), None))
+    let (id, _name) = kh
+        .spawn_agent(minimal_manifest(), None)
+        .await
         .expect("spawn failed");
 
     let agents = kh.list_agents();
